This is how far I've come with Uppgift 1, enkelriktad association.
I've created the classes for Human, Dog, and PetGUI. Right now I'm in
the middle of trying to figure out how to make the GUI work. 

Wondering if I have to create an ArrayList to store Humans and Dogs?
It's not in the UML. Or will it work ok just to have one dog and human
at a time stored in variables? Also I'm not sure exactly how enkelriktad
association should look like in the code.